+++ CHANGELOG.md
@@ -93,3 +93,4 @@
 - 2023-09-13: Wire qualifier, smtp, onboarding, mod, +2
 - 2023-09-14: Rate chat, prompts, mod, pools, +2
 - 2023-09-15: Valid mod, session, mod, scheduler, +6
+- 2023-09-15: Valid prompt, chat, scheduler, mod, +5